refactor: 完成所有指令重构
